Slight changes
I was trying to make the instructions clear.  I'll add a few comments to certain paragraphs too.

Also, I think we need to state the requirements for this repo. For example, it could be done like this:

## Package requirements:
```
pytorch == 1.10.2
numpy == 1.20.3
scikit-learn == 0.22.2.post1
tqdm == 4.62.3
pandas == 1.0.3
```

Or as Huggingface accomplished this: This repository is tested on Python 3.6+, Flax 0.3.2+, PyTorch 1.3.1+ and TensorFlow 2.3+.